uasdk: force unitTest to use vendor SSL libraries
- try to avoid DLL hell by copying the vendor-supplied
  SSL libraries next to the test binaries

+++ devOpcuaSup/open62541/DataElementOpen62541.cpp
@@ -136,7 +136,7 @@ DataElementOpen62541::createMap (const UA_DataType *type,
                 if (timefrom) {
 #ifdef UA_ENABLE_TYPEDESCRIPTION
                     if (*timefrom == member->memberName) {
-                        timesrc = memberOffs;
+                        timesrc = static_cast<int>(memberOffs);
                     }
 #endif
                 }
